add Main.java, fix FileReader.java & FileWriter.java

+++ src/main/java/homeworks/modul3/FileWriter.java
@@ -1,6 +1,7 @@
 package homeworks.modul3;
 
 import java.nio.file.Files;
+import java.nio.file.StandardOpenOption;
 import java.nio.file.Paths;
 import java.io.IOException;
 
@@ -16,11 +17,16 @@ class FileWriter implements FileProcessor {
     @Override
     public void process() throws MyFileException {
         try {
-            Files.writeString(Paths.get(filePath), content);
+            if (Files.exists(Paths.get(filePath))) {
+                if (Files.size(Paths.get(filePath)) > 0) {
+                    Files.writeString(Paths.get(filePath), "\n" + content, StandardOpenOption.APPEND);
+                }
+            } else {
+                Files.writeString(Paths.get(filePath), content);
+            }
             System.out.println("Данные успешно записаны в файл");
         } catch (IOException e) {
             throw new MyFileException("Ошибка при записи в файл: " + e.getMessage());
         }
     }
 }
-
